fix: wxWidgets format string assertion - use %d instead of %zu for size_t
Fixed assertion failure in wxArgNormalizer when using wxString::Format
with %zu format specifier on MinGW-w64.

Changes:
- ProxyListPanel.cpp:92 - use %d with static_cast<int>() for row numbers
- SubscriptionPanel.cpp:98 - use %d with static_cast<int>() for proxy counts

This ensures compatibility with wxWidgets strict format checking.

+++ src/ui/ProxyListPanel.cpp
@@ -89,7 +89,7 @@ void ProxyListPanel::loadProxies(const std::string& subId) {
     store_->DeleteAllItems();
     for (const auto& p : proxies_) {
         wxVector<wxVariant> row;
-        row.push_back(wxVariant(wxString::Format("%zu", store_->GetCount() + 1)));
+        row.push_back(wxVariant(wxString::Format("%d", static_cast<int>(store_->GetCount() + 1))));
         row.push_back(wxVariant(p.address));
         row.push_back(wxVariant(p.port));
         row.push_back(wxVariant(delayMap.count(p.indexid) ? delayMap[p.indexid] : "-"));
